feat: Support product bundling with packs
Introduce the `isPack` and `packItems` fields to the `Product` type to enable the creation and management of product bundles. This change enhances inventory management by allowing users to define products as collections of other products.

Also, add a `resetDemoProductsOnly` function to the `StorageService` and update the `SuperAdminView` to default to the demo products tab for easier testing. Icons for editing and packs have been added to `lucide-react`.